Refactor getBnsIdentity function
