+++ packages/bierzo-wallet/src/routes/account/update/index.tsx
@@ -1,5 +1,4 @@
-import { ChainId, Fee, Identity, TransactionId } from "@iov/bcp";
-import { BnsConnection } from "@iov/bns";
+import { Fee, Identity, TransactionId } from "@iov/bcp";
 import React from "react";
 import * as ReactRedux from "react-redux";
 
@@ -9,10 +8,8 @@ import { generateRegisterUsernameTxWithFee } from "../../../communication/reques
 import { ChainAddressPairWithName } from "../../../components/AddressesTable";
 import PageMenu from "../../../components/PageMenu";
 import { getConfig, SupportedChain } from "../../../config";
-import { getConnectionForChainId } from "../../../logic/connection";
-import { ExtendedIdentity } from "../../../store/identities";
 import { RootState } from "../../../store/reducers";
-import { getChainAddressPairWithNamesSorted } from "../../../utils/tokens";
+import { getBnsIdentity, getChainAddressPairWithNamesSorted } from "../../../utils/tokens";
 import { TRANSACTIONS_ROUTE } from "../../paths";
 import ConfirmRegistration from "./components/ConfirmUpdate";
 import IovnameAccountUpdate from "./components/IovnameForm";
@@ -20,15 +17,6 @@ import NameAccountUpdate from "./components/NameForm";
 
 function onSeeTrasactions(): void {
   history.push(TRANSACTIONS_ROUTE);
-}
-
-export function getBnsIdentity(identities: ReadonlyMap<ChainId, ExtendedIdentity>): Identity | undefined {
-  for (const identity of Array.from(identities.values()).map(ext => ext.identity)) {
-    if (getConnectionForChainId(identity.chainId) instanceof BnsConnection) {
-      return identity;
-    }
-  }
-  return undefined;
 }
 
 async function getPersonalizedAddressRegistrationFee(
